+++ mockup/patterns/thememapper/js/lessbuilderview.js
@@ -13,8 +13,8 @@ define([
       '<div class="buttonBox">' +
         '<label for="lessFileName">Save as:</label>' +
         '<input id="lessFileName" type="text" placeholder="filename" />' +
-        '<a id="compileBtn" class="context" href="#">Compile</a>' +
-        '<a id="errorBtn" class="btn btn-default" href="#">Clear</a>' +
+        '<button id="compileBtn" class="btn btn-primary">Compile</button>' +
+        '<button id="errorBtn" class="btn btn-default">Clear</button>' +
       '</div>' +
     '</div>'
   );
@@ -52,10 +52,22 @@ define([
     },
     setFilename: function() {
         var self = this;
+
         if( self.app.lessPaths['save'] === undefined ) {
             return;
         }
-        var f = self.app.lessPaths['save'];
+
+        var filePath = self.app.lessPaths['less'];
+        var devPath = self.app.devPath[0];
+        var prodPath = self.app.prodPath[0];
+
+        if( filePath == devPath ) {
+            var f = prodPath;
+        }
+        else {
+            var f = self.app.lessPaths['save'];
+        }
+
         f = f.substr(f.lastIndexOf('/') + 1, f.length);
         self.$filename.attr('placeholder', f);
     },